Extract QusestionsAccordion component to coach common
